Add support for user accounts and authentication.
Create pydantic types and SQLAlchemy ORMs to represent a user in the runtime and database, respectively. Integrate fastapi-users login route. Add authentication dependency to upload route.  Add a script to retry database connection. Add a script to seed database with users. Add shell script to demonstrate composition of a postgresql service with the application.